Rewrite of doc-comment parser in a more flexible form.
The old code was hardcoded to understand specific '@-tags', and would be
extended by adding clauses into a big case-when statatement.

The new archiecture provides generic parsing of JavaDoc @-tags, and allows
handler objects to be registered against tag names to process the information
contained in the tag.

Parsing of doc comments is now handled earlier, too.  We used to just collect
the comment text when the AS file was parsed, and then actually parse the
comment when generating the documentation.  Now, parsing of doc-comments
happens as they're encountered, which is far more natural.

The code is probably slightly more buggy than that which was replaced.
To make up for that, this commit also adds support for JavaDoc {@link ...}
syntax, which the old implementation didn't understand.


git-svn-id: http://svn.badgers-in-foil.co.uk/as2api/trunk/as2api@75 4f4d65cf-00e9-0310-82ac-fa7a581d2dae