Hack-in initial support for parsing types outside the set being documented.
This allows us to put the MM 'Classes' directory into the classpath, and have
references to Object, String et al get resolved, without actually having to
generate pages documenting those types.  The main utility of this
functionality is in showing what members are inherited from the external type.

This hopefully means an end to lots of bogus
  "Found no definition of type known locally as <blah>"
messages.

The implementation sucks a bit, and there are still places that will attempt
to hyperlink to pages which don't exist.


git-svn-id: http://svn.badgers-in-foil.co.uk/as2api/trunk/as2api@127 4f4d65cf-00e9-0310-82ac-fa7a581d2dae