Return of inheritance diagrams as a permanent feature.
By adding --draw-diagrams to the command line options, each package index page
will include a diagram showing inheritance relationships between interfaces and
a diagram showing inheritance relationships between classes.  The classes and
interfaces in the diagrams are hyperlinks to the 'detail' pages for the
respective types.

The diagram generation requires 'dot', a part of the Graphviz package, and
by default it is assumed that 'dot' is available in the system search PATH.
If this is not the case, the program file to be used can be specified with the
new --dot-exe option, e.g.

 ... --dot-exe /home/dave/graphviz/bin/dot ...

 Error handling isn't all that great at the moment (if 'dot' is missing, the
 diagrams are absent, but as2api is silent).


git-svn-id: http://svn.badgers-in-foil.co.uk/as2api/trunk/as2api@135 4f4d65cf-00e9-0310-82ac-fa7a581d2dae