A small lexer optimisation.
Rather than having the lexer scan for every possible keyword token, simply
match all keywords / identifiers with the same rule, and have the lexical
action classify the input when the rule is matched.

This is a seedup because at any point in the input, the number of different
rules we have to test to recognise the next token is quite a lot smaller.
The lexical action can then try to lookup the matched text in a hashtable,
which should be faster that the previous linear search through all the
keyword.

My completely trivial test case went from about 37 seconds to 34 with this
change.


git-svn-id: http://svn.badgers-in-foil.co.uk/as2api/trunk/as2api@207 4f4d65cf-00e9-0310-82ac-fa7a581d2dae