Very early support for documenting the changes between two API versions. This is inspired by <http://jdiff.sf.net/>, but the output and implementation are nowhere near as nice yet.
To use, the new --oldrev-classpath command line option should a an exact copy
of the --classpath option, but pointing at the location of the 'old' revision
of the classes.  Both versions of the classes will be parsed, and some
additional HTML files will appear in the 'changes' subdirectory of the
documentation.


git-svn-id: http://svn.badgers-in-foil.co.uk/as2api/trunk/as2api@218 4f4d65cf-00e9-0310-82ac-fa7a581d2dae