Slurp entire .as file into a StringScanner at once, rather than streaming it line-by-line.
This simplifies the code a bit, and speeds up lexing.  Handling of #include
lines has been moved into the lexer (rather than beeing a preprocessing step)
to make this possible.

This change has the potential to degrade performance if the files to be loaded
are very 'large' (for some unspecified definition), but that seems unlikely to
be the case most of the time.  I can't spot a 'streaming' alternative to
StringScanner.

If StringScanner is compatable with the Ruby mmap wrapper, making use of mmap,
when available, might well provide a further boost, and would make the lack of
streaming a non-issue.


git-svn-id: http://svn.badgers-in-foil.co.uk/as2api/trunk/as2api@377 4f4d65cf-00e9-0310-82ac-fa7a581d2dae